Turing Tumble components can now be inserted
Each pin and compSlot can have different types of components inserted
into them, when it gets a component it will pass this piece to the child
component which can then access and modify this piece in this child
component. Every turing tumble piece now has it's own Angular Component
that will be able to display it's own UI logic later on.

Closes #6