Fix placement of components
Placing a component will now give a new unique component to that slot
fixing the previous bug where it would give the same piece to multiple
slots. The placement logic has been put into the compSlot and Pin
components so that it can be given the text of the component it will be
making and then creates a new component if it is given a new component
to place. But if given the component it already holds then it will
'click' this component instead, so ramp will change direction for
example

Closes #9