Logic of all components added to app, marbles now follow game Logic
All board components have been added such that when marble is dropped
it follows the board and releases marbles when needed. The gear and
gearBit uses a graph style deep first search to find the components it
connects to then decide if it will switch the components out. If the
direction needs changed it will create a new piece instead of switching
the direction of the piece to stay in line with the angular hook Logic

Closes #11