Puzzle board can now be made
The make puzzle page allows users to make a puzzle following a simple
instruction based system. It helps set the amount of pieces available
during a puzzle and what the output should be. Propmt card now made.
Plus board service has been removed in favour of Output() in the child
components that access the board.