Can now create a puzzle and upload it to firebase
Firebase was chosen as an easy to use and free backend so it can show
off the back end capabilities for the site. After going through the make
puzzle page users enter title, description and difficulty.

Data had to be converted to JSON to upload and then parsed back into its
types when it gets downloaded.